requirements.txt
django==4.2.11
python-dotenv==1.0.0
djangorestframework==3.14.0
markdown==3.5.1
django-cors-headers==4.3.1
requests==2.31.0
pillow==8.4.0
django-cache-memoize==0.2.0
asgiref==3.6.0
attrs==23.1.0
coverage==7.2.7
drf-spectacular==0.26.2
gunicorn==21.2.0
inflection==0.5.1
jsonschema==4.17.3
packaging==23.1
pyrsistent==0.19.3
pytz==2023.3
PyYAML==6.0
tzdata==2023.3
uritemplate==4.1.1
